cast to float just in case

+++ collect/collect.go
@@ -378,7 +378,7 @@ func (i *InMemCollector) collect() {
 					return
 				}
 				i.processSpan(sp)
-				i.Metrics.Gauge("collector_collect_loop_duration_ns", time.Now().Sub(startTime))
+				i.Metrics.Gauge("collector_collect_loop_duration_ns", float64(time.Now().Sub(startTime)))
 				continue
 			case sp, ok := <-i.fromPeer:
 				if !ok {
@@ -386,14 +386,14 @@ func (i *InMemCollector) collect() {
 					return
 				}
 				i.processSpan(sp)
-				i.Metrics.Gauge("collector_collect_loop_duration_ns", time.Now().Sub(startTime))
+				i.Metrics.Gauge("collector_collect_loop_duration_ns", float64(time.Now().Sub(startTime)))
 				continue
 			case <-i.reload:
 				i.reloadConfigs()
 			}
 		}
 
-		i.Metrics.Gauge("collector_collect_loop_duration_ns", time.Now().Sub(startTime))
+		i.Metrics.Gauge("collector_collect_loop_duration_ns", float64(time.Now().Sub(startTime)))
 	}
 }
